+++ inc/woocommerce/hooks.php
@@ -22,9 +22,26 @@ remove_action( 'woocommerce_before_main_content', 	'woocommerce_breadcrumb', 			
 remove_action( 'woocommerce_before_main_content', 	'woocommerce_output_content_wrapper', 		10 );
 remove_action( 'woocommerce_after_main_content', 	'woocommerce_output_content_wrapper_end', 	10 );
 remove_action( 'woocommerce_sidebar', 				'woocommerce_get_sidebar', 					10 );
+remove_action( 'woocommerce_after_shop_loop', 		'woocommerce_pagination', 					10 );
+remove_action( 'woocommerce_before_shop_loop', 		'woocommerce_result_count', 				20 );
+remove_action( 'woocommerce_before_shop_loop', 		'woocommerce_catalog_ordering', 			30 );
 add_action( 'woocommerce_before_main_content', 		'storefront_before_content', 				10 );
 add_action( 'woocommerce_after_main_content', 		'storefront_after_content', 				10 );
 add_action( 'storefront_content_top', 				'woocommerce_breadcrumb', 					10 );
+
+add_action( 'woocommerce_after_shop_loop',			'storefront_sorting_wrapper',				9 );
+add_action( 'woocommerce_after_shop_loop', 			'woocommerce_catalog_ordering', 			10 );
+add_action( 'woocommerce_after_shop_loop', 			'woocommerce_result_count', 				20 );
+add_action( 'woocommerce_after_shop_loop', 			'woocommerce_pagination', 					30 );
+add_action( 'woocommerce_after_shop_loop',			'storefront_sorting_wrapper_close',			31 );
+
+add_action( 'woocommerce_before_shop_loop',			'storefront_sorting_wrapper',				19 );
+add_action( 'woocommerce_before_shop_loop', 		'woocommerce_catalog_ordering', 			20 );
+add_action( 'woocommerce_before_shop_loop', 		'woocommerce_result_count', 				30 );
+add_action( 'woocommerce_before_shop_loop', 		'woocommerce_pagination', 					40 );
+add_action( 'woocommerce_before_shop_loop',			'storefront_sorting_wrapper_close',			41 );
+
+
 
 /**
  * Products